Unravel contracts: separate rain functions
