+++ src/contract/contract.h
@@ -2,4 +2,3 @@
 
 bool VerifyCPIDSignature(std::string sCPID, std::string sBlockHash, std::string sSignature);
 bool SignBlockWithCPID(const std::string& sCPID, const std::string& sBlockHash, std::string& sSignature, std::string& sError, bool bAdvertising=false);
-std::string executeRain(std::string sRecipients);
