+++ src/contract/rain.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+std::string executeRain(std::string sRecipients);
